refactor: move SerializableRuleCore to rule_core mod
