+++ crates/config/src/rule_core.rs
@@ -1,12 +1,17 @@
-use serde::{Deserialize, Serialize};
-
+use crate::fixer::{FixerError, SerializableFixer};
 use crate::rule::referent_rule::RuleRegistration;
 use crate::rule::Rule;
+use crate::rule::{RuleSerializeError, SerializableRule};
 use crate::transform::{apply_env_transform, Transformation};
+use crate::DeserializeEnv;
+use crate::GlobalRules;
+
 use ast_grep_core::language::Language;
 use ast_grep_core::matcher::{KindMatcher, KindMatcherError, RegexMatcher, RegexMatcherError};
 use ast_grep_core::meta_var::{MetaVarEnv, MetaVarMatcher, MetaVarMatchers};
 use ast_grep_core::{Doc, Matcher, Node, Pattern, PatternError, StrDoc};
+use serde::{Deserialize, Serialize};
+use serde_yaml::Error as YamlError;
 
 use bit_set::BitSet;
 use schemars::JsonSchema;
@@ -58,6 +63,74 @@ pub fn try_deserialize_matchers<L: Language>(
     map.insert(key, try_from_serializable(matcher, lang.clone())?);
   }
   Ok(map)
+}
+
+#[derive(Debug, Error)]
+pub enum RuleConfigError {
+  #[error("Fail to parse yaml as RuleConfig")]
+  Yaml(#[from] YamlError),
+  #[error("Rule is not configured correctly.")]
+  Rule(#[from] RuleSerializeError),
+  #[error("fix pattern is invalid.")]
+  Fixer(#[from] FixerError),
+  #[error("constraints is not configured correctly.")]
+  Constraints(#[from] SerializeConstraintsError),
+}
+
+type RResult<T> = std::result::Result<T, RuleConfigError>;
+
+/// Used for global rules, rewriters, and pyo3/napi
+#[derive(Serialize, Deserialize, Clone, JsonSchema)]
+pub struct SerializableRuleCore<L: Language> {
+  /// Specify the language to parse and the file extension to include in matching.
+  pub language: L,
+  /// A rule object to find matching AST nodes
+  pub rule: SerializableRule,
+  /// Additional meta variables pattern to filter matching
+  pub constraints: Option<HashMap<String, SerializableMetaVarMatcher>>,
+  /// Utility rules that can be used in `matches`
+  pub utils: Option<HashMap<String, SerializableRule>>,
+  /// A dictionary for metavariable manipulation. Dict key is the new variable name.
+  /// Dict value is a [transformation] that specifies how meta var is processed.
+  /// See [transformation doc](https://ast-grep.github.io/reference/yaml/transformation.html).
+  pub transform: Option<HashMap<String, Transformation>>,
+  /// A pattern string or a FixConfig object to auto fix the issue.
+  /// It can reference metavariables appeared in rule.
+  /// See details in fix [object reference](https://ast-grep.github.io/reference/yaml/fix.html#fixconfig).
+  pub fix: Option<SerializableFixer>,
+}
+
+impl<L: Language> SerializableRuleCore<L> {
+  pub fn get_deserialize_env(&self, globals: &GlobalRules<L>) -> RResult<DeserializeEnv<L>> {
+    let env = DeserializeEnv::new(self.language.clone()).with_globals(globals);
+    if let Some(utils) = &self.utils {
+      let env = env.register_local_utils(utils)?;
+      Ok(env)
+    } else {
+      Ok(env)
+    }
+  }
+
+  fn get_meta_var_matchers(&self) -> RResult<MetaVarMatchers<StrDoc<L>>> {
+    Ok(if let Some(constraints) = self.constraints.clone() {
+      try_deserialize_matchers(constraints, self.language.clone())?
+    } else {
+      MetaVarMatchers::default()
+    })
+  }
+
+  pub fn get_matcher(&self, globals: &GlobalRules<L>) -> RResult<RuleWithConstraint<L>> {
+    let env = self.get_deserialize_env(globals)?;
+    let rule = env.deserialize_rule(self.rule.clone())?;
+    let matchers = self.get_meta_var_matchers()?;
+    let transform = self.transform.clone();
+    Ok(
+      RuleWithConstraint::new(rule)
+        .with_matchers(matchers)
+        .with_utils(env.registration)
+        .with_transform(transform),
+    )
+  }
 }
 
 pub struct RuleWithConstraint<L: Language> {
